Refaktoroi oppiaineenvalintakenttä omaan tiedostoon
